program to brute force caesar
As we know that that caesar cipher is a very weak cipher technique. It can be easily brute forced and the original text can be retrived even without knowing the key. The total no of keys that caesar cipher can have is only 25 and hence we have to check only 25 assumptions.